# nginx/docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Recipeblog API'
  version: ''
paths:
...
  /api/users/me/avatar/:
    put:
      operationId: Добавление аватара
      description: Добавление аватара текущего пользователя
      parameters: []
      security:
        - Token: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAvatar'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetAvatarResponse'
          description: 'Аватар успешно добавлен'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
        - Пользователи
    delete:
      operationId: Удаление аватара
      description: Удаление аватара текущего пользователя
      parameters: []
      security:
        - Token: []
      responses:
        '204':
          description: Аватар успешно удален
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
        - Пользователи
...
components:
  schemas:
...
    SetAvatar:
      description: 'Добавление аватара пользователя'
      type: object
      properties:
        avatar:
          description: 'Картинка, закодированная в Base64'
          example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABAgMAAABieywaAAAACVBMVEUAAAD///9fX1/S0ecCAAAACXBIWXMAAA7EAAAOxAGVKw4bAAAACklEQVQImWNoAAAAggCByxOyYQAAAABJRU5ErkJggg=='
          type: string
          format: binary
      required:
        - avatar
    SetAvatarResponse:
      type: object
      properties:
        avatar:
          type: string
          format: uri
          description: 'Ссылка на аватар'
          example: 'http://foodgram.example.org/media/users/image.png'
...
    ValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        field_name:
          description: 'Название поля, в котором произошли ошибки. Таких полей может быть несколько'
          example: [ 'Обязательное поле.' ]
          type: array
          items:
            type: string
...
    AuthenticationError:
      description: Пользователь не авторизован
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
...
  responses:
    ValidationError:
      description: 'Ошибки валидации в стандартном формате DRF'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    AuthenticationError:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
...
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token